Resolved Multifaceted Abstraction in EthereumUtilTest by Extract Class Technique
